refactor: Comprehensive BigDecimal migration and package restructuring
CHANGES:
- WalletManager: Moved from com.bidding.engine to com.bidding.shared (correct package placement)
  * Converted all API methods to use BigDecimal instead of double
  * registerNewWallet() now uses BigDecimal.ZERO for initialization
  * execute(), depositDirectly(), transferMoney() now accept BigDecimal parameters
  * Maintained double overloads for backward compatibility

- Balance.java: Full BigDecimal conversion with synchronized operations
  * All balance fields (currentBalance, lockedBalance) now use BigDecimal
  * Operations: lockAmount(), unlockAmount(), commitLockedAmount()
  * Proper thread-safety with synchronized blocks

- Item.java: Added BigDecimal support for pricing
  * firstprice now uses BigDecimal type
  * Constructor overloads for both double and BigDecimal inputs

- ItemManager.java: Fixed updateItemPriceByAdmin() method
  * Added BigDecimal import
  * Line 115: Convert double newPrice to BigDecimal.valueOf(newPrice)
  * Ensures type compatibility with Item.setFirstprice(BigDecimal)

- DepositMethod.java: Fixed BOM character encoding issue
  * Removed UTF-8 BOM (\\ufeff) that caused javac parsing errors
  * Ensured proper UTF-8 no-BOM encoding for compilation

- AuctionRoom.java: Maintained thread-safe observer pattern
  * Uses CopyOnWriteArrayList for concurrent access
  * ConcurrentHashMap-backed Set for participant management

VERIFICATION:
- All 12 Java source files compile successfully
- Zero compilation errors across the entire codebase
- Correct package structure: com.bidding.shared (9 files), com.bidding.engine (2 files)
- Ready for Main.java implementation

+++ Biddingweb/Biddingweb/demo/src/main/java/com/bidding/shared/ItemManager.java
@@ -1,5 +1,6 @@
 package com.bidding.shared;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -109,7 +110,7 @@ public class ItemManager {
         // 2. Nếu đúng là Admin, tìm và cập nhật giá
         Item item = getItemById(itemId);
         if (item != null) {
-            item.setFirstprice(newPrice);
+            item.setFirstprice(BigDecimal.valueOf(newPrice));
             System.out.println("Cập nhật giá mặt hàng thành công!");
         } else {
             System.out.println("Lỗi: Không tìm thấy sản phẩm này.");
